+++ src/utils/check_in.hpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 
-bool check_in(vector<int> result, int j) {
+bool check_in(vector<float> result, float j) {
   for (auto i : result) {
     if (i == j) return true;
   }
