+++ src/utils/length.hpp
@@ -3,7 +3,16 @@
 using namespace std;
 
 
-int length(vector<int> solution, vector<vector<int> > distance_matrix) {
+/**
+ * @brief Length of tour
+ * 
+ * This function computes the length of a given tour.
+ * 
+ * @param solution The tour
+ * @param distance_matrix The distance matrix
+ * @return The length of the tour
+ */
+int length(vector<float> solution, vector<vector<float> > distance_matrix) {
   int length = 0;
   int starting_node = solution[0];
   int from_node = starting_node;
